backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "louis", email = "louisjoejordan@gmail.com"},
    {name = "leander", email = "l.maerkisch@gmail.com"}
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "python-dotenv>=1.0.1",
    "websockets>=13.0.1",
    "openai>=1.46.0",
    "mistralai>=1.1.0",
    "instructor>=1.5.1",
    "requests>=2.32.3",
    "pytest>=8.3.3",
    "ruff>=0.6.9",
    "pip>=24.2",
    "python-multipart>=0.0.12",
    "pytest-asyncio>=0.24.0",
    "certifi>=2024.8.30",
    "httpx>=0.27.2",
]
requires-python = ">=3.12"
readme = "README.md"

[tool.setuptools]
packages = ["src"]

[project.scripts]
start = "src.main:start"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# Added pytest configuration
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
